requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
ccxt==4.4.35
openai>=1.0.0
google-genai>=0.2.0
pandas>=1.5.0
numpy>=1.24.0
httpx>=0.23.0
requests>=2.28.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0
